chore(dev): move dev logic into its own folder
