+++ src/dev/Dev.vue
@@ -2,8 +2,8 @@
 
 import fullSprite from '@rei/cedar-icons/dist/all-icons.svg';
 import SinkWrapper from './SinkWrapper.vue'; // eslint-disable-line
-import { CdrLink } from './index';
-import cedarcss from './css/main.scss';
+import { CdrLink } from '../index';
+import cedarcss from '../css/main.scss';
 import routes from './router';
 
 export default {
